+++ lib/src/promise.dart
@@ -1,0 +1,6 @@
+part of babylon;
+
+@JS()
+class Promise<T> {
+  external Promise<dynamic> then(T onFulfilled, dynamic onRejected);
+}