+++ lib/src/shadowlight.dart
@@ -2,7 +2,7 @@ part of babylon;
 
 @JS()
 class ShadowLight extends Light {
-  void Function(Matrix viewMatrix,List<AbstractMesh> renderList,Matrix result) customProjectionMatrixBuilder;
+  void Function(Matrix viewMatrix, List<AbstractMesh> renderList, Matrix result) customProjectionMatrixBuilder;
   Vector3 transformedPosition;
   Vector3 transformedDirection;
   external Vector3 get position;